// web/widget/gql/schema.ts
export type PickArrayType<T> = T extends (infer U)[] ? U : NonNullable<T>
export const crs = { type: 'name', properties: { name: 'urn:ogc:def:crs:EPSG::4326' } }
export type Maybe<T> = T | null | undefined
// ...
export type FindBusinessQuery = {
  findBusinessList?: Maybe<
    Array<{
      id: string
      name: string
      description?: Maybe<string>
      photoUrl?: Maybe<string>
      markerUrl?: Maybe<string>
      ownerId: string
      address?: Maybe<{ [key: string]: unknown }>
      location: { longitude: number; latitude: number }
    }>
  >
}